refactor: ProductFixture 생성
