+++ src/test/java/kitchenpos/OrderAcceptanceTest.java
@@ -13,6 +13,7 @@ import org.junit.jupiter.api.Test;
 import java.math.BigDecimal;
 import java.util.List;
 
+import static kitchenpos.fixture.ProductFixture.스키야키;
 import static kitchenpos.step.MenuGroupStep.메뉴_그룹_생성_요청하고_아이디_반환;
 import static kitchenpos.step.MenuStep.메뉴_생성_요청하고_아이디_반환;
 import static kitchenpos.step.OrderStep.주문_상태_변경_요청;
@@ -20,7 +21,6 @@ import static kitchenpos.step.OrderStep.주문_생성_요청;
 import static kitchenpos.step.OrderStep.주문_생성_요청하고_주문_반환;
 import static kitchenpos.step.OrderStep.주문_조회_요청;
 import static kitchenpos.step.ProductStep.상품_생성_요청하고_아이디_반환;
-import static kitchenpos.step.ProductStep.스키야키;
 import static kitchenpos.step.TableStep.테이블_생성_요청하고_테이블_반환;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
